# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/colecciones": {
      "post": {
        "operationId": "ColeccionController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearColeccionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coleccion"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una colección propia (usuario con sesión)",
        "tags": [
          "Colecciones"
        ]
      },
      "get": {
        "operationId": "ColeccionController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Coleccion"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar mis colecciones (usuario con sesión)",
        "tags": [
          "Colecciones"
        ]
      }
    },
    "/colecciones/{id}": {
      "get": {
        "operationId": "ColeccionController_consultar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coleccion"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver una colección (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      },
      "patch": {
        "operationId": "ColeccionController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarColeccionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coleccion"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modificar una colección; juegoIds reemplaza la lista (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      },
      "delete": {
        "operationId": "ColeccionController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una colección (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      }
    },
    "/colecciones/{id}/juegos": {
      "post": {
        "operationId": "ColeccionController_agregarJuego",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarJuegoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coleccion"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: el dato ya existe (duplicado) o está en uso y no se puede borrar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CUU Administrar colección: agregar un juego (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      }
    },
    "/colecciones/{id}/juegos/{juegoId}": {
      "delete": {
        "operationId": "ColeccionController_quitarJuego",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "juegoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Datos inválidos: el mensaje indica qué campo falló, un id de la URL que no es un número, o un id del body que no existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión: falta el token, es inválido o está vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso: la acción es solo para ADMIN o el recurso es de otro usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          },
          "404": {
            "description": "No existe el recurso pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuestaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "CUU Administrar colección: quitar un juego (el dueño o un ADMIN)",
        "tags": [
          "Colecciones"
        ]
      }
    },
...
  },
  "info": {
    "title": "DGame (MyGameSearcher) API",
    "description": "API REST que recomienda de 1 a 3 videojuegos según géneros, características y plataforma.\n\n**Cómo autenticarse:** hacer `POST /auth/login` con email y contraseña, copiar el `token` de la respuesta y pegarlo en el botón **Authorize** (arriba a la derecha). Desde ahí, todas las pruebas se mandan con `Authorization: Bearer <token>`.\n\n**Niveles de acceso:** público (sin token), usuario con sesión iniciada, y ADMIN. El resumen de cada ruta indica quién puede usarla.\n\n**Errores:** todas las respuestas de error tienen la forma `{ \"statusCode\", \"message\", \"error\" }`, con el mensaje en español. 400: datos inválidos · 401: sin sesión o token inválido · 403: sin permiso · 404: no existe · 409: conflicto (duplicado o en uso).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token que devuelve POST /auth/login"
      }
    },
    "schemas": {
...
      "Genero": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
      "ErrorRespuestaDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "description": "Mensaje en español. En los errores de validación (400) es una lista con un mensaje por campo",
            "oneOf": [
              {
                "type": "string",
                "example": "No se encontró el género con id 9999"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "El nombre es obligatorio"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "Plataforma": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "Caracteristica": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "ClasificacionEdad": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
...
      "Juego": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "titulo": {
            "type": "string"
          },
          "anioLanzamiento": {
            "type": "number"
          },
          "descripcion": {
            "type": "string"
          },
          "imagenUrl": {
            "type": "string",
            "nullable": true
          },
          "clasificacionEdad": {
            "$ref": "#/components/schemas/ClasificacionEdad"
          },
          "plataformas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plataforma"
            }
          },
          "generos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genero"
            }
          },
          "caracteristicas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Caracteristica"
            }
          }
        },
        "required": [
          "id",
          "titulo",
          "anioLanzamiento",
          "descripcion",
          "imagenUrl",
          "clasificacionEdad",
          "plataformas",
          "generos",
          "caracteristicas"
        ]
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "rol": {
            "enum": [
              "USUARIO",
              "ADMIN"
            ],
            "type": "string"
          },
          "fechaRegistro": {
            "format": "date-time",
            "type": "string"
          },
          "plataforma": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Plataforma"
              }
            ]
          }
        },
        "required": [
          "id",
          "nombre",
          "apellido",
          "email",
          "rol",
          "fechaRegistro",
          "plataforma"
        ]
      },
...
      "CrearColeccionDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100
          },
          "descripcion": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "juegoIds": {
            "nullable": true,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "nombre"
        ]
      },
      "Coleccion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "format": "date-time",
            "type": "string"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "juegos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Juego"
            }
          }
        },
        "required": [
          "id",
          "nombre",
          "descripcion",
          "fechaCreacion",
          "usuario",
          "juegos"
        ]
      },
      "ActualizarColeccionDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100
          },
          "descripcion": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "juegoIds": {
            "nullable": true,
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "AgregarJuegoDto": {
        "type": "object",
        "properties": {
          "juegoId": {
            "type": "number"
          }
        },
        "required": [
          "juegoId"
        ]
      },
...
    }
  }
}