# src/coleccion/dto/actualizar-coleccion.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CrearColeccionDto } from './crear-coleccion.dto';

// El dueño no se puede cambiar: sale del token al crear la colección
export class ActualizarColeccionDto extends PartialType(CrearColeccionDto) {}